% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attorneygeneral}
\alias{attorneygeneral}
\title{Ballot Orders for Attorney General}
\format{
A data frame with 254 rows and 4 variables:
\describe{
\item{County}{name of Texas county}
\item{Branch}{candidate's ballot position}
\item{Paxton}{candidate's ballot position}
\item{Smitherman}{candidate's ballot position}
}
}
\source{
\url{mailto:dgrant@shsu.edu}
}
\usage{
attorneygeneral
}
\description{
Ballot orders for 2014 Texas Republican Primary
election, attorney general race.
}
\keyword{datasets}
