% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comptroller}
\alias{comptroller}
\title{Ballot Orders for Comptroller}
\format{
A data frame with 254 rows and 4 variables:
\describe{
\item{County}{name of Texas county}
\item{Hegar}{candidate's ballot position}
\item{Hilderbran}{candidate's ballot position}
\item{Medina}{candidate's ballot position}
\item{Torres}{candidate's ballot position}
}
}
\source{
\url{mailto:dgrant@shsu.edu}
}
\usage{
comptroller
}
\description{
Ballot orders for 2014 Texas Republican Primary
election, comptroller race.
}
\keyword{datasets}
