% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_compatibility.R
\name{expand}
\alias{expand}
\title{Expand a partially ordered list}
\usage{
expand(vectors)
}
\arguments{
\item{vectors}{list}
}
\value{
matrix
}
\description{
Compute all vectors of m elements that are compatible with
a given partial ordering of the m elements.
}
