% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lieutenantgovernor}
\alias{lieutenantgovernor}
\title{Ballot Orders for Lieutenant Governor}
\format{
A data frame with 254 rows and 4 variables:
\describe{
\item{County}{name of Texas county}
\item{Dewhurst}{candidate's ballot position}
\item{Patrick}{candidate's ballot position}
\item{Patterson}{candidate's ballot position}
\item{Staples}{candidate's ballot position}
}
}
\source{
\url{mailto:dgrant@shsu.edu}
}
\usage{
lieutenantgovernor
}
\description{
Ballot orders for 2014 Texas Republican Primary
election, lieutenant governor race.
}
\keyword{datasets}
