% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_concordance.R
\name{linear_concordance}
\alias{linear_concordance}
\title{Execute Linear Concordance Test}
\usage{
linear_concordance(.data, scores, direction = "uni")
}
\arguments{
\item{.data}{data.frame}

\item{scores}{double}

\item{direction}{character}
}
\value{
double
}
\description{
Given a set of orderings and a user-specified score vector,
execute either the uni- or bi-directional LC test
}
