% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{railroadcommissioner}
\alias{railroadcommissioner}
\title{Ballot Orders for Railroad Commissioner}
\format{
A data frame with 254 rows and 4 variables:
\describe{
\item{County}{name of Texas county}
\item{Berger}{candidate's ballot position}
\item{Boyuls}{candidate's ballot position}
\item{Christian}{candidate's ballot position}
\item{Sitton}{candidate's ballot position}
}
}
\source{
\url{mailto:dgrant@shsu.edu}
}
\usage{
railroadcommissioner
}
\description{
Ballot orders for 2014 Texas Republican Primary
election, railroad commissioner race.
}
\keyword{datasets}
