% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_compatibility.R
\name{rank_compatibility}
\alias{rank_compatibility}
\title{Execute Rank Compatibility test}
\usage{
rank_compatibility(.data, alternatives, cushion = 1)
}
\arguments{
\item{.data}{data.frame}

\item{cushion}{double}

\item{alternative}{list}
}
\value{
list
}
\description{
Given a set of orderings, an alternative preference
criterion, and a cushion for the number of biased
agents, compute the p value for the RC test.
}
