% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_concordance.R
\name{rank_concordance}
\alias{rank_concordance}
\title{Execute Rank Concordance test}
\usage{
rank_concordance(.data, alternative, cushion = 1)
}
\arguments{
\item{.data}{data.frame}

\item{alternative}{list}

\item{cushion}{double}
}
\value{
double
}
\description{
Given a set of orderings, an alternative preference
criterion, and a cushion for the number of biased
agents, compute the p value for the RC test.
}
