% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{senator}
\alias{senator}
\title{Ballot Orders for Senator}
\format{
A data frame with 254 rows and 4 variables:
\describe{
\item{County}{name of Texas county}
\item{Cleaver}{candidate's ballot position}
\item{Cope}{candidate's ballot position}
\item{Cornyn}{candidate's ballot position}
\item{Mapp}{candidate's ballot position}
\item{Reasor}{candidate's ballot position}
\item{Stockman}{candidate's ballot position}
\item{Stovall}{candidate's ballot position}
\item{Vega}{candidate's ballot position}
}
}
\source{
\url{mailto:dgrant@shsu.edu}
}
\usage{
senator
}
\description{
Ballot orders for 2014 Texas Republican Primary
election, senator race.
}
\keyword{datasets}
